Build an example JSON parameter file

Function used to write JSON parameter file. A function is provided instead than a json file to ensure directories to match the user folder tree.

build_example_param_file(
  json_path = tempfile(fileext = "_sen2r_params.json"),
  overwrite = TRUE
)

Arguments

json_path
Path of the output file. Default is to save it on a temporary file, whose path is returned.

overwrite
Logical value: should existing output file be overwritten? (default: TRUE)

Value

The path of the created file.

Note

License: GPL 3.0

References

L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020). "sen2r": An R toolbox for automatically downloading and preprocessing Sentinel-2 satellite data. Computers & Geosciences, 139, 104473. DOI: 10.1016/j.cageo.2020.104473, URL: http://sen2r.ranghetti.info/.

Examples

build_example_param_file()

#> [1] "/tmp/RtmpMw5dY5/file74fb44cdcb1d_sen2r_params.json"