Check GUI dependencies

Check if suggested dependencies required to run the sen2r GUI are present.

check_gui_deps(abort = TRUE)

Arguments

abort
Logical parameter: if TRUE (default), the function aborts in case some packages need to be installed; if FALSE, a warning is shown.

Value

Logical (invisible): TRUE if all the required packages are installed, FALSE (if abort = FALSE) or an error if some is missing.

Details

Packages required to run the GUI (shiny- and leaflet-related packages), which were sen2r dependencies until version 1.3.4, are now suggested dependencies. sen2r can be used without them with the exception of the GUI.

Note

License: GPL 3.0

References

L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020). "sen2r": An R toolbox for automatically downloading and preprocessing Sentinel-2 satellite data. Computers & Geosciences, 139, 104473. DOI: 10.1016/j.cageo.2020.104473, URL: http://sen2r.ranghetti.info/.

Examples

# \donttest{
check_gui_deps()
# }