Find, download and preprocess Sentinel-2 images

The function is a wrapper to perform the entire processing chain to find, download and pre-process Sentinel-2 data. Input is a set of parameters that can be passed with a list or file (parameter param_list) or singularly (see the descriptions of all the other parameters).

fidolasen_s2(param_list = NULL, sel_sensor = NA, online = NA,
  overwrite_safe = NA, rm_safe = NA, step_atmcorr = NA,
  steps_reqout = NA, timewindow = NA, timeperiod = NA, extent = NA,
  s2tiles_selected = NA, list_prods = NA, reference_path = NA, res = NA,
  unit = NA, proj = NA, resampling = NA, outformat = NA,
  compression = NA, overwrite = NA, path_l1c = NA, path_l2a = NA,
  path_tiles = NA, path_out = NA, path_tiles_subdirs = NA,
  path_out_subdirs = NA)

Arguments

param_list
List of input parameters: it can be both an R list or the path of a JSON file. If some parameters are passed both as elements of param_list and as function arguments, the values passed as function arguments are considered. If some parameters are missing in param_list and are not provided as arguments, default values will be used. Use the function s2_gui() to create a complete list of parameters.

sel_sensor
=NA,

online
=NA,

overwrite_safe
=NA,

rm_safe
=NA,

step_atmcorr
=NA,

steps_reqout
=NA,

timewindow
=NA,

timeperiod
=NA,

extent
=NA,

s2tiles_selected
=NA,

list_prods
=NA,

reference_path
=NA,

res
=NA,

unit
=NA,

proj
=NA,

resampling
=NA,

outformat
=NA,

compression
=NA,

overwrite
=NA,

path_l1c
=NA,

path_l2a
=NA,

path_tiles
=NA,

path_out
=NA,

path_tiles_subdirs
=NA,

path_out_subdirs
=NA

timewindow
=NA,