Get information from S2 short name

This accessory function extracts metadata included in the name of a Sentinel-2 product which follows the sen2r naming convention (see s2_shortname).

fs2nc_getElements(s2_names, format = "list", abort = TRUE)

Arguments

s2_names
A vector of Sentinel-2 product names in the sen2r naming convention.

format
One between list of data.frame.

abort
Logical parameter: if TRUE (default), the function aborts in case any of s2_names is not recognised; if FALSE, a warning is shown, and a list with only the element "type"='unrecognised' is returned.

Value

A list or a data.frame of the output metadata.

Note

License: GPL 3.0

Examples

# Define product name
fs2nc_examplename <-
  "/path/of/the/product/S2A1C_20170603_022_32TQQ_TOA_20.tif"

# Return metadata
fs2nc_getElements(fs2nc_examplename)

#> $type
#> [1] "tile"
#> 
#> $mission
#> [1] "A"
#> 
#> $level
#> [1] "1C"
#> 
#> $sensing_date
#> [1] "2017-06-03"
#> 
#> $id_orbit
#> [1] "022"
#> 
#> $id_tile
#> [1] "32TQQ"
#> 
#> $prod_type
#> [1] "TOA"
#> 
#> $res
#> [1] "20m"
#> 
#> $file_ext
#> [1] "tif"
#>