Preprocess (clip, reproject, reshape, mask) SAFE products

The function is a simplified wrapper of sen2r() to preprocess existing SAFE products. No online research is performed, neither sen2cor is applied to existing L1C products. Output products are GeoTIFF files clipped to a desired

geograbber_process(
  path_safe,
  path_out,
  bbox = NA,
  list_prods = c("BOA"),
  res = NA,
  ...
)

Arguments

path_safe
Path of the directory in which SAFE products are searched.

path_out
Path of the directory in which Sentinel-2 output products are generated.

bbox
(optional) Four-length numeric vector with the bounding box of the output products (xmin, ymin, xmax, ymax), in geographic coordinates.

list_prods
(optional) Character vector with the values of the products to be processed (accepted values: "TOA", "BOA", "SCL", "TCI"). Default is "BOA".

res
(optional) Numeric vector of length 2 with the x-y resolution for output products. NA (default) means that the resolution is kept as native.

...
Other additional sen2r arguments: extent_name, list_indices, index_source, mask_type, max_mask, mask_smooth, mask_buffer, clip_on_extent, extent_as_mask, reference_path, res_s2, unit, proj, resampling, resampling_scl, outformat, index_datatype, compression, overwrite, path_tiles, path_merged, path_indices, path_subdirs, thumbnails, parallel, tmpdir, rmtmp. See sen2r documentation for details.