Manage GIPP parameters for Sen2Cor

read_gipp() reads Ground Image Processing Parameters (GIPP) from an XML file.

set_gipp() modifies values of a list of GIPP in an XML file.

reset_gipp() restores default GIPP values (using values set in the L2A_GIPP.xml default Sen2Cor file).

read_gipp(gipp_names, gipp_path = NA)

set_gipp(gipp = list(), gipp_path = NA, use_dem = NA)

reset_gipp(gipp_path = NA)

Arguments

gipp_names
Character vector with the names of the parameters to be read.

gipp_path
Character path of the output GIPP XML file. If NA (default), the default sen2r GIPP path is used.

gipp
Ground Image Processing Parameters (GIPP). It is possible to specify both the path of an existing XML file or a list of parameters in the form parameter_name = "value", where parameter_name is the name of the parameter as specified in the L2A_GIPP.xml file of the used Sen2Cor version (case insensitive), and "value" is the character value which the user wants to set (notice that, in the case the user wants to specify the value NONE, both "NONE" and NA can be used, but not NULL, which has the effect to maintain the value specified in the XML file). Elements whose name is missing in the XML file are skipped.

use_dem
Logical: if TRUE, a DEM is set for being used for topographic corection in the XML specified with argument gipp_path; if FALSE, no DEM is set for being used; if NA (default), the existing option is maintained. Notice that, in case use_dem = TRUE and a DEM directory is already set in the XML file, no changes are performed; conversely; if no directories were specified, a subdirectory "srtm90" of the default sen2r directory is set. To set another directory, use argument gipp in the form gipp = list(DEM_Directory = tempdir(), ...) (replacing tempdir() with the desired path).

Value

read_gipp() returns a named list of GIPP with the required parameters (values not found in the XML are skipped).

set_gipp() and reset_gipp() return NULL (functions are called for their side effects).

Note

License: GPL 3.0

Examples

# \donttest{
# Read default values
read_gipp(c("dem_directory", "dem_reference"))

#> $dem_directory
#> [1] "/home/lranghetti/.sen2r/srtm90"
#> 
#> $dem_reference
#> [1] "NONE"
#> 

# Edit one value and save the output as temporary file
set_gipp(list(DEM_Directory = NA), gipp_path = gipp_temp <- tempfile())
# Read the parameters in the created temporary files
read_gipp(c("DEM_Directory", "DEM_Reference"), gipp_path = gipp_temp)

#> $DEM_Directory
#> [1] "NONE"
#> 
#> $DEM_Reference
#> [1] "NONE"
#> 

# Read the original (not edited) XML file
read_gipp(c("DEM_Directory", "DEM_Reference"))

#> $DEM_Directory
#> [1] "/home/lranghetti/.sen2r/srtm90"
#> 
#> $DEM_Reference
#> [1] "NONE"
#> 

# Reset to default Sen2Cor GIPP values
reset_gipp(gipp_path = gipp_temp)

#> Default Sen2Cor parameters were written in file /home/lranghetti/.sen2r/sen2r_L2A_GIPP.xml, and will be used with sen2cor() and sen2r() unless different parameters will be specified.
#> IMPORTANT NOTE: for backward compatibility, the parameter "DEM_Directory" was maitained to its default value. This does not grant homogeneity between Level-2A SAFE products generated locally with Sen2Cor and downloaded from ESA Hub (which make use of DEM for topographic correction).To grant it, activate topographic correction using:
#>  > set_gipp(use_dem = TRUE)
#> In a future sen2r release, this will be the default Sen2Cor behaviour.

# Read again values
read_gipp(c("DEM_Directory", "DEM_Reference"), gipp_path = gipp_temp)

#> $DEM_Directory
#> [1] "NONE"
#> 
#> $DEM_Reference
#> [1] "http://data_public:GDdci@data.cgiar-csi.org/srtm/tiles/GeoTIFF/"
#> 

# }