Download and install aria2.

This function download and install standalone version of aria2 for Windows.

install_aria2(aria2_dir = system.file(package = "sen2r"), force = FALSE)

Arguments

aria2_dir
(optional) Path where aria2 executable will be installed (default: the package path).

force
(optional) Logical: if TRUE, install even if it is already installed (default is FALSE).

Value

The path of aria2

Note

License: GPL 3.0

Examples

# NOT RUN {
install_aria2()
# }