Return the mountpoint of the input directory (if it is mounted)

The function checks if the input directory is a subdirectory of a mountpoint of a certain protocol. At the moment, it works only on unix operating systems.

mountpoint(path, protocol = NA)

Arguments

path
The path to be checked

protocol
(optional) Vector of protocol types. If NA (default), all the protocols are considered.

Value

The path of the parent mountpoint for mounted directories; if the input directory is not mounted, NULL is returned. NULL is returned also if the operating system is not unix (together with a warning message). An attribute "protocol" contains the protocol of the mountpoint.

Note

License: GPL 3.0