Compute maps of spectral indices

Create maps of a set of spectral indices. Since gdal_calc.py is used to perform computations, output files are physical rasters (no output VRT is allowed).

s2_calcindices(infiles, indices, outdir = ".", parameters = NULL,
  format = NA, subdirs = NA, tmpdir = NA, compress = "DEFLATE",
  dataType = "Int16")

Arguments

infiles
A vector of input filenames. Input files are paths of BOA (or TOA) products already converted from SAFE format to a format managed by GDAL (use s2_translate to do it); their names must be in the fidolasen-S2 naming convention (s2_shortname).

indices
Character vector with the names of the required indices. Values should be included in names corresponding to the Abbreviations of the following indices: IDB FIXME the list of the accepted values is a subset; this reference will be replaced with an internal html page integrated in the shiny interface).

outdir
(optional) Full name of the output directory where the files should be created (default: current directory). outdir can bot be an existing or non-existing directory (in the second case, its parent directory must exists). If it is a relative path, it is expanded from the common parent directory of infiles.

parameters
(optional) Values of index parameters. This variable must be a named list, in which each element is a list of parameters, i.e.: parameters = list("SAVI" = list("a" = 0.5)) TODO: add default values (in the json)

format
(optional) Format of the output file (in a format recognised by GDAL). Default is the same format of input images (or "GTiff" in case of VRT input images).

subdirs
(optional) Logical: if TRUE, different indices are placed in separated outfile subdirectories; if FALSE, they are placed in outfile directory; if NA (default), subdirectories are created only if more than a single spectral index is required.

tmpdir
(optional) Path where intermediate VRT will be created. Default is in a hidden subdirectory (called .vrt) of the common parent directory of infiles. Set tmpdir=tempdir() if you do not want to keep the intermediate files after reboot.

compress
(optional) In the case a GTiff format is present, the compression indicated with this parameter is used.

dataType
(optional) Numeric datatype of the ouptut rasters: if "Float32" or "Float64" is chosen, numeric values are not rescaled; if "Int16" (default) or "UInt16", values are multiplicated by a 10000 scale factor.

Value

A vector with the names of the created products.

Note

License: GPL 3.0