Return the Dates Of Passage over some orbits

The function allows to know which Sentinel-2 passages should pass over certain orbits during a defined time interval. Dates are intended to be in UTC time. Notice that this is the expected calendar: some unexpected events (e.g. technical problems, or early working phases during first stages of acquisition) could cause the data unavailability even if an acquisition was expected. Notice also that some orbits (030, 073 and 116) acquire across UTC midnight: in this cases, the date is assumed to be the one of the acquisition after midnight (which corresponds to the date in local time).

s2_dop(s2_orbits = 1:143, timewindow = "10 days", mission = c("2A", "2B"))

Arguments

s2_orbits

A vector of Sentinel-2 orbits (as integer numbers or 3-length character). Default is all the 143 orbits.

timewindow

Temporal window for querying: Date object of length 1 (single day) or 2 (time window). Is it possible to pass also integer (or difftime) values, which are interpreted as the next n days (if positive) or the past n days (if negative). Also strings which can be interpreted as time ranges are accepted (see examples). Default is the next 10 days (one cycle).

mission

(optional) Vector with the desired Sentinel-2 missions ("2A", "2B" or both). Default is both.

Value

A data table with the dates (column "date"), the missions (column "mission") and the orbits (column "orbit"). An empty data table with the same structure is returned if no passages were found with the passed settings.

Note

License: GPL 3.0

References

L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020). "sen2r": An R toolbox for automatically downloading and preprocessing Sentinel-2 satellite data. Computers & Geosciences, 139, 104473. doi: 10.1016/j.cageo.2020.104473 , URL: https://sen2r.ranghetti.info/.

Author

Luigi Ranghetti, phD (2019) luigi@ranghetti.info

Examples

# All the passages in a cycle of 10 days over all the orbits
s2_dop()
#>            date mission orbit
#>   1: 2022-04-05      2A   116
#>   2: 2022-04-05      2A   117
#>   3: 2022-04-05      2A   118
#>   4: 2022-04-05      2A   119
#>   5: 2022-04-05      2A   120
#>  ---                         
#> 282: 2022-04-14      2B   040
#> 283: 2022-04-14      2B   041
#> 284: 2022-04-14      2B   042
#> 285: 2022-04-14      2B   043
#> 286: 2022-04-14      2B   044

# The passages in the current month over two orbits
s2_dop(c("022", "065"), "this month")
#>           date mission orbit
#>  1: 2022-04-01      2A   065
#>  2: 2022-04-03      2B   022
#>  3: 2022-04-06      2B   065
#>  4: 2022-04-08      2A   022
#>  5: 2022-04-11      2A   065
#>  6: 2022-04-13      2B   022
#>  7: 2022-04-16      2B   065
#>  8: 2022-04-18      2A   022
#>  9: 2022-04-21      2A   065
#> 10: 2022-04-23      2B   022
#> 11: 2022-04-26      2B   065
#> 12: 2022-04-28      2A   022

# The dates in which Sentinel-2A will pass in next six weeks over one orbit
s2_dop("022", "6 weeks", mission = "2A")$date
#> [1] "2022-04-08" "2022-04-18" "2022-04-28" "2022-05-08"

# The date in which Sentinel-2A would be passed in the last 10 days over one orbit
s2_dop("022", "-10 days", mission = "2A")$date
#> [1] "2022-03-29"

# All the orbits covered today
s2_dop(timewindow = Sys.Date(), mission = "2B")$orbit
#>  [1] "045" "046" "047" "048" "049" "050" "051" "052" "053" "054" "055" "056"
#> [13] "057" "058"

# The passages in a fixed time window for one orbit
s2_dop(65, as.Date(c("2018-08-01", "2018-08-31")))
#>          date mission orbit
#> 1: 2018-08-05      2B   065
#> 2: 2018-08-10      2A   065
#> 3: 2018-08-15      2B   065
#> 4: 2018-08-20      2A   065
#> 5: 2018-08-25      2B   065
#> 6: 2018-08-30      2A   065

# A research with no passages found
s2_dop(22, "2018-08-16", mission = "2A")
#> Empty data.table (0 rows and 3 cols): date,mission,orbit