Merge S2 tiles with the same date and orbit

The function merge the input Sentinel-2 files with the same date, orbit number, product type and file format. Outputs are a set of products in the same format of corresponding input files.

s2_merge(infiles, outdir = ".", subdirs = NA, tmpdir = NA, format = NA,
  compress = "DEFLATE", out_crs = "", overwrite = FALSE)

Arguments

infiles
A vector of input filenames. Input files are paths of products already converted from SAFE format to a format managed by GDAL (use s2_translate to do it); their names must be in the fidolasen-S2 naming convention (s2_shortname).

outdir
(optional) Full name of the output directory where the files should be created (default: current directory). outdir can bot be an existing or non-existing directory (in the second case, its parent directory must exists). If it is a relative path, it is expanded from the common parent directory of infiles.

subdirs
(optional) Logical: if TRUE, differet output products are placed in separated outfile subdirectories; if FALSE, they are placed in outfile directory; if NA (default), subdirectories are created only if infiles relate to more than a single product.

tmpdir
(optional) Path where intermediate VRT will be created. Default is in a hidden subdirectory (called .vrt) of the common parent directory of infiles. Set tmpdir=tempdir() if you do not want to keep the intermediate files after reboot.

format
(optional) Format of the output file (in a format recognised by GDAL). Default is to maintain each input format.

compress
(optional) In the case a GTiff format is present, the compression indicated with this parameter is used.

out_crs
(optional) proj4string (character) of the output CRS (default: the CRS of the first input file). The tiles with CRS different from out_crs will be reprojected (and a warning returned).

overwrite
Logical value: should existing output files be overwritten? (default: FALSE)

Value

A vector with the names of the merged products (just created or already existing).

Note

License: GPL 3.0