Create RGB images from S2 reflectance products.

Function to create RGB images from Sentinel-2 reflectances.

s2_rgb(infiles, prod_type = NA, rgb_bands = 4:2, scaleRange = NA,
  outdir = NA, subdirs = NA, format = NA, compress = "DEFLATE",
  tmpdir = NA, rmtmp = TRUE, overwrite = FALSE)

Arguments

infiles
A vector of input filenames. Input files are paths of products already converted from SAFE format to a format managed by GDAL (use s2_translate to do it); their names must be in the sen2r naming convention (safe_shortname).

prod_type
(optional) Output product (see safe_shortname for the list of accepted products). If not provided, it is retrieved from the file name.

rgb_bands
(optional) 3-length integer vector, which the number of the bands to be used respectively for red, green and blue. Default is 4:2 (true colours). It is also possible to pass a list of 3-length integer vectors in order to create multiple RGB types for each input file. # TODO Notice that this is the actual number name of the bands: so, to use i.e. BOA band 11 (1610nm) use the number 11, even if band 11 is the 10th band of a BOA product (because band 10 is missing).

scaleRange
(optional) Range of valid values. If can be a 2-length integer vector (min-max for all the 3 bands) or a 6-length vector or 3x2 matrix (min red, min green, min blue, max red, max green, max blue). Default is c(0,2500) for true colours RGB, c(0,7500) for others.

outdir
(optional) Full name of the existing output directory where the files should be created. Default is the same directory of input reflectance files. # FIXME use a subdir with product name

subdirs
(optional) Logical: if TRUE, different indices are placed in separated outfile subdirectories; if FALSE, they are placed in outfile directory; if NA (default), subdirectories are created only if more than a single spectral index is required.

format
(optional) Format of the output file (in a format recognised by GDAL). Default is the same format of input images (or "GTiff" in case of VRT input images).

compress
(optional) In the case a GTiff format is present, the compression indicated with this parameter is used.

tmpdir
(optional) Path where intermediate files (VRT) will be created. Default is a temporary directory.

rmtmp
(optional) Logical: should temporary files be removed? (Default: TRUE)

overwrite
(optional) Logical value: should existing thumbnails be overwritten? (default: TRUE)

Value

A vector with the names of the created images.

Note

License: GPL 3.0