Correct L1C products using sen2cor

The function uses sen2cor from the docker to manually correct L1C products.

s2_sen2cor(l1c_prodlist = NULL, l1c_dir = NULL, outdir = NULL,
  n_procs = 1, overwrite = FALSE)

Arguments

l1c_prodlist
List of L1C product names to be corrected. They can be both product names with full/relative path or only names of SAFE products (in this case, also l1c_dir argument must be provided). SAFE products must be unzipped. Note that, at this stage, all products must be in the same directory (this will be fixed).

l1c_dir
Full or relative path of input L1C products.

outdir
Directory where output L2A products will be placed.

n_procs
Number of processors (integer) to use (default is 1, single processor).

overwrite
Logical value: should existing output L2A products be overwritten? (default: FALSE)

Value

Vector character with the list ot the output products (being corrected or already existing)

Note

License: GPL 3.0

Examples

# NOT RUN {
pos <- sp::SpatialPoints(data.frame("x"=12.0,"y"=44.8), proj4string=sp::CRS("+init=epsg:4326"))
time_window <- as.Date(c("2017-05-01","2017-07-30"))
example_s2_list <- s2_list(spatial_extent=pos, tile="32TQQ", time_interval=time_window)
s2_download(example_s2_list, outdir=tempdir())
s2_sen2cor(names(example_s2_list)[1], l1c_dir=tempdir(), outdir=tempdir())
# }