Rename products using a shorten convention

This function renames a Sentinel-2 product in order to obtain shorten names. See the details for the structure of the adopted schema. The function applies only to product names (not to single granule names), since it is thought to be applied to entire products.

s2_shortname(prod_name, ext = NULL, full.name = TRUE, set.seed = NA)

Arguments

prod_name
Input Sentinel-2 product name: it can be both a complete path or a simple file name. If the file exists, the product is scanned to retrieve the tiles list and UTM zones; otherwise the basename is simply translated in he short naming convention.

ext
(optional) Extension of the output filename (default: none)

full.name
Logical value: if TRUE (default), all the input path is maintained (if existing); if FALSE, only basename is returned.

set.seed
(internal parameter) Logical value: if TRUE, the generation of random tile elements is univocal (this is useful not to assign different names to the same element); if FALSE, the generation is completely random (this is useful not to assign the same name to products with the same names (e.g. product with old name downloaded in different moments which contains different subsets of tiles). Default value is TRUE if the file exists locally, FALSE if not.

Value

Output product name

Details

ESA Sentinel-2 naming convention is particularly long-winded; moreover, the convention changed after December 6th 2016, causing products to follow two different schemes.

The convention here adopted follows this schema: S2mll_yyyymmdd_rrr_ttttt

where:

S2mll shows the mission ID (S2A or S2B) and the product level (1C or 2A);

yyyymmdd is the sensing date (e.g. 20170603 for 2017-06-03) ; the hour is skipped, since a single sensor can not pass two times in a day on the same tile);

rrr is the relative orbit number (e.g. 022);

ttttt is the tile number (e.g. 32TQQ).

Products which contains only one tile (date > 2016-12-05 or < 2016-12-06 but containing only one tile) follows entirely this convention; for products with more than one tile (or that can not be checked), the field ttttt is replaced by:

uuxxx if the UTM zone is unique (where uu is the UTM zone, and xxx is a random numeric sequence), e.g. 32876;

yyxxx if the UTM zone is not unique or it it can not be checked (being yy a random literal A-Z sequence, and xxx a random numeric sequence).

Since the tile number is composed by two numeric and three literal character, these two conventions (one composed by five numeric and the other by two literal and three numeric characters) ensure that these three cases can be distinguished.

Note

License: GPL 3.0

Examples

# Define product names
s2_compactname_example <- file.path(
  "/non/existing/path",
  "S2A_MSIL1C_20170603T101031_N0205_R022_T32TQQ_20170603T101026.SAFE")
s2_oldname_example <-
  "S2A_OPER_PRD_MSIL1C_PDMC_20160809T051732_R022_V20150704T101337_20150704T101337.SAFE"
s2_existing_example <- "/replace/with/the/main/path/of/existing/product/with/oldname/convention"

# With compact names, it works also without scanning the file
s2_shortname(s2_compactname_example, ext="tif")

#> [1] "/non/existing/path/S2A1C_20170603_022_32TQQ.tif"

# With old names, without scanning the file the id_tile is not retrieved,
# so the tile filed is replaced with a random sequence
unlist(lapply(rep(s2_oldname_example,5), s2_shortname, full.name=FALSE))

#> [1] "S2A1C_20150704_022_CV600" "S2A1C_20150704_022_EA466"
#> [3] "S2A1C_20150704_022_MH732" "S2A1C_20150704_022_UW174"
#> [5] "S2A1C_20150704_022_AI402"

# NOT RUN {
# If the file exists locally, the tile is retrieved from the content
# (if unique; if not, a random sequence is equally used, but it is
# always the same for the same product)
s2_shortname(s2_existing_example, full.name=FALSE)
# }