Create thumbnails for S2 products.

Function to create thumbnail images for Sentinel-2 products. BOA and TOA multiband images are rendered as false colour JPEG images; SCL maps are rendered as 8-bit PNG; other singleband images (like spectral indices) are rendered as JPEG images with a standard color palette. Some improvements still have to be done:

allowing the possibility to chose the palette and the limits (for now, the range -1 to 1 is always used);

adding support for RGB products.

Output images are georeferenced.

s2_thumbnails(infiles, prod_type = NA, rgb_type = "SwirNirR", dim = 1024,
  outdir = NA, tmpdir = NA, overwrite = FALSE)

Arguments

infiles
A vector of input filenames. Input files are paths of products already converted from SAFE format to a format managed by GDAL (use s2_translate to do it); their names must be in the fidolasen-S2 naming convention (s2_shortname).

prod_type
(optional) Output product (see s2_shortname for the list of accepted products). If not provided, it is retrieved from the file name.

rgb_type
(optional) For BOA and TOA products, this value determine the type of false colours to be used for the thumbnails:

"SwirNirR" (default) for SWIR-NIR-Red;

"NirRG" for NIR-Red-Green;

"RGB" for true colours;

dim
Integer value, with the maximum greater dimension in pixels (width or height) of the output images (default: 1024 px). If this is lower than the corresponding dimension of the maps, maps are rescaled before producing the thumbnails; otherwise the original dimensions are maintained. To keep the original size in any case, set dim = Inf.

outdir
(optional) Full name of the existing output directory where the files should be created. Default is a subdirectory (named "thumbnails") of the parent directory of each input file.

tmpdir
(optional) Path where intermediate VRT will be created. Default is a temporary directory.

overwrite
(optional) Logical value: should existing thumbnails be overwritten? (default: TRUE)

Value

A vector with the names of the created images.

Note

License: GPL 3.0