Correct L1C products using Sen2Cor

The function uses Sen2Cor to manually correct L1C products. Standalone version of sen2cor (version 2.8.0 or 2.5.5) is used.

sen2cor(
  l1c_prodlist = NULL,
  l1c_dir = NULL,
  outdir = NULL,
  proc_dir = NA,
  tmpdir = NA,
  rmtmp = TRUE,
  gipp = NULL,
  use_dem = NA,
  tiles = NULL,
  parallel = FALSE,
  overwrite = FALSE,
  .log_message = NA,
  .log_output = NA
)

Arguments

l1c_prodlist
List of L1C product names to be corrected. They can be both product names with full/relative path or only names of SAFE products (in this case, also l1c_dir argument must be provided). SAFE products must be unzipped. Note that, at this stage, all products must be in the same directory (this will be fixed).

l1c_dir
Full or relative path of input L1C products. If NULL (default), l1c_prodlist must already be a vector of full paths.

outdir
Directory where output L2A products will be placed. If NULL (default), each product is left in the parent directory of l1c_prodlist.

proc_dir
(optional) Directory where processing is applied. If NA (default), processing is done in l1c_dir and output L2A product is then moved to outdir, unless l1c_dir is a subdirectory of a SAMBA mountpoint under Linux: in this case, L1C input products are copied in a temporary directory (specified with argument tmpdir), processing is done there and then L2A is moved to outdir. This is required under Linux systems when l1c_dir is a subdirectory of a unit mounted with SAMBA, otherwise Sen2Cor would produce empty L2A products.

tmpdir
(optional) Path where processing is performed if a temporary working directory is required (see argument proc_dir). Be sure tmpdir not to be a SAMBA mountpoint under Linux. Default is a temporary directory. If tmpdir is a non-empty folder, a random subdirectory will be used.

rmtmp
(optional) Logical: should temporary files be removed? (Default: TRUE)

gipp
(optional) Ground Image Processing Parameters (GIPP) to be passed to Sen2Cor (see set_gipp() for further details). Note: this argument takes effect only in the current execution of sen2cor() function; to permanently change GIPP values, use function set_gipp().

use_dem
(optional) Logical: if TRUE, Sen2Cor is set to use a Digital Elevation Model for topographic correction (reflecting what is done for Level-2A SAFE images provided by ESA Hub); if FALSE, it is set not to perform topographic correction (reflecting the current default Sen2Cor behaviour); if NA (default), the option set in the user's L2A_GIPP.xml Sen2Cor configuration file is respected (in case the user never edited it, the current default setting is not to perform topographic correction).

Notes:

if TRUE, the path used to read or store DEM files and the online source used to download missing DEM tiles are respectively the DEM_Directory and DEM_Reference parameters set in the default sen2r GIPP XML file (the user can read them with the function read_gipp(c("DEM_Directory", "DEM_Reference"))). In case one or both these parameters were set to "NONE", a subdirectory "srtm90" of the default sen2r directory is used as DEM directory, and/or the CGIAR SRTM 90m is set as online source. To set another directory or reference, use argument gipp in the form gipp = list(DEM_Directory = tempdir(), DEM_Reference ="another_reference", ...) (replacing tempdir() with the desired path and specifying the online resource).

Currently the default value is NA in order to grant backward compatibility. In a future release of sen2r, the default value will be set to TRUE, so to grant homogeneity between Level-2A products downloaded from ESA Hub and generated using Sen2Cor.

tiles
Vector of Sentinel-2 Tile strings (5-length character) to be processed (default: process all the tiles found in the input L1C products).

parallel
(optional) Logical: if TRUE, Sen2Cor instances are launched in parallel using multiple cores; if FALSE (default), they are launched in series on a single core. The number of cores is automatically determined; specifying it is also possible (e.g. parallel = 4).

overwrite
Logical value: should existing output L2A products be overwritten? (default: FALSE)

.log_message
(optional) Internal parameter (it is used when the function is called by sen2r()).

.log_output
(optional) Internal parameter (it is used when the function is called by sen2r()).

Value

Vector character with the list ot the output products (being corrected or already existing)

Note

License: GPL 3.0

Examples

if (FALSE) {
# Download an L1C SAFE product
example_s2_list <- s2_list(
  spatial_extent = st_sfc(st_point(c(12.0, 44.8)), crs=st_crs(4326)),
  tile = "32TQQ",
  time_interval = as.Date(c("2017-05-01","2017-07-30"))
)
s2_download(example_s2_list, outdir = tempdir())

# Correct it applying a topographic correction
sen2cor(
  names(example_s2_list)[1],
  l1c_dir = tempdir(),
  outdir = tempdir(),
  use_dem = TRUE
)
}