Find, download and preprocess Sentinel-2 images

The function is a wrapper to perform the entire processing chain to find, download and pre-process Sentinel-2 data. Input is a set of parameters that can be passed with a list or file (parameter param_list) or singularly (see the descriptions of all the other parameters).

sen2r(param_list = NULL, gui = NA, preprocess = TRUE,
  s2_levels = c("l1c", "l2a"), sel_sensor = c("s2a", "s2b"),
  online = TRUE, apihub = NA, downloader = "wget",
  overwrite_safe = FALSE, rm_safe = "no", step_atmcorr = "auto",
  max_cloud_safe = 100, timewindow = NA, timeperiod = "full",
  extent = NA, extent_name = "sen2r", s2tiles_selected = NA,
  s2orbits_selected = NA, list_prods = c("BOA"), list_rgb = NA,
  list_indices = NA, index_source = "BOA", rgb_ranges = NA,
  mask_type = NA, max_mask = 100, mask_smooth = 0, mask_buffer = 0,
  clip_on_extent = TRUE, extent_as_mask = FALSE, reference_path = NA,
  res = NA, res_s2 = "10m", unit = "Meter", proj = NA,
  resampling = "near", resampling_scl = "near", outformat = "GTiff",
  rgb_outformat = "GTiff", index_datatype = "Int16",
  compression = "DEFLATE", rgb_compression = "90", overwrite = FALSE,
  path_l1c = NA, path_l2a = NA, path_tiles = NA, path_merged = NA,
  path_out = NA, path_rgb = NA, path_indices = NA,
  path_subdirs = TRUE, thumbnails = TRUE, parallel = TRUE,
  processing_order = "by_date", use_python = TRUE, tmpdir = NA,
  rmtmp = TRUE, log = NA)

Arguments

param_list
(optional) List of input parameters: it can be both an R list or the path of a JSON file. If some parameters are passed both as elements of param_list and as function arguments, the values passed as function arguments are considered. If some parameters are missing in param_list and are not provided as arguments, default values will be used. Use the function s2_gui() to create a complete list of parameters. If param_list is NULL (default), values given with the parameters below (or default values for parameters not provided) are used.

gui
(optional) Logical: if TRUE, function s2_gui() is launched before starting to process in order to set or load parameters; if FALSE, the function uses parameters passed with param_list or with other function arguments. Default is FALSE if param_list is not NULL, TRUE elsewhere.

preprocess
(optional) Logical: TRUE (default) to perform also preprocessing steps, FALSE not to (do only find, download and atmospheric correction).

s2_levels
(optional) Character vector of length 1 or 2, with Sentinel-2 levels required for processing steps or as output. This parameter is used only if preprocess = FALSE (otherwise, the required levels are derived from list_prods). Accepted values: "l1c" and "l2a"; default: "l2a".

sel_sensor
(optional) Character vector of length 1 or 2, with Sentinel-2 sensors to be used. Accepted values: "s2a" and "s2b"; default: c("s2a","s2b").

online
(optional) Logical: TRUE (default) to search for available products on SciHub (and download if needed); FALSE to work only with already downloaded SAFE products.

apihub
Path of the text file containing credentials of scihub account. If NA (default) the default credentials (username "user", password "user") will be used.

downloader
(optional) Character value corresponding to the executable which should be used to download SAFE products. It could be one among "wget" (default) and "aria2". If aria2 is not installed, Wget will be used instead.

overwrite_safe
(optional) Logical: TRUE to overwrite existing products with products found online or manually corrected, FALSE (default) to skip download and atmospheric correction for products already existing.

rm_safe
(optional) Character: should SAFE products be deleted after preprocessing? "yes" means to delete all SAFE; "no" (default) not to delete; "l1c" to delete only Level-1C products.

step_atmcorr
(optional) Character vector to determine how to obtain Level-2A SAFE products:

"auto" (default) means that L2A is first searched on SciHub: if found, it is dowloaded, if not, the corresponding Level-1C is downloaded and sen2cor is used to produce L2A;

"scihub" means that sen2cor is always used from L1C products downloaded from SciHub;

"l2a" means that they are downloaded if available on SciHub, otherwise they are skipped (sen2cor is never used);

"no" means that L2A are not considered (processing chain makes use only of L1C products).

max_cloud_safe
(optional) Integer number (0-100) containing the maximum cloud level of each SAFE to be considered (default: no filter). It it used to limit the research of SAFE products to "good" images, so it is applied only to non-existing archives (existing SAFE are always used). In thish sense, this parameter is different from max_mask, which can be used to set a maximum cloud coverage over output extents. Notice also that this value is used to filter on the basis of the metadata "Cloud cover percentage" associated to each SAFE, so it is not based on the cloud mask defined with the processing options.

timewindow
(optional) Temporal window for querying: Date object of length 1 (single day) or 2 (time window). Default is NA, meaning that no filters are used if online = FALSE, and all found images are processed; if online = TRUE, last 90 days are processed. Is it possible to pass also integer (or difftime) values, which are interpreted as the last n days.

timeperiod
(optional) Character:

"full" (default) means that all the images included in the time window are considered;

"seasonal" means that only the single seasonal periods in the window are used (i.e., with a time window from 2015-06-01 to 2017-08-31, the periods 2015-06-01 to 2015-08-31, 2016-06-01 to 2016-08-31 and 2017-06-01 to 2017-08-31 are considered).

extent
(optional) Spatial extent on which to clip products (it can be both the path of a vector file or a geoJSON). Default is NA for offline mode (meaning no extent: all found tiles are entirely used); in online mode, a sample extent is used as default.

extent_name
(optional) Name of the area set as extent, to be used in the output file names. Default is "sen2r" The name is an alphanumeric string which cannot contain points nor underscores, and that cannot be a five-length string with the same structure of a tile ID (two numeric and three uppercase character values).

s2tiles_selected
(optional) Character vector with the Sentinel-2 tiles to be considered (default is NA, meaning all the tiles).

s2orbits_selected
(optional) Character vector with the Sentinel-2 orbits to be considered (still to be implemented; for now, all the accepted values are listed).

list_prods
(optional) Character vector with the values of the products to be processed (accepted values: "TOA", "BOA", "SCL", "TCI"). Default is "BOA".

list_rgb
(optional) Character vector with the values of the RGB images to be produced. Images are in the form RGBrgbx, when:

x is B (if source is BOA) or T (is source is TOA);

r g and b are the the number of the bands to be used respectively for red, green and blue, in hexadecimal format. Notice that this is the actual number name of the bands: so, to use i.e. BOA band 11 (1610nm) use the value "b", even if band 11 is the 10th band of a BOA product (because band 10 is missing). Default is no one (NA).

list_indices
(optional) Character vector with the values of the spectral indices to be computed. Default is no one (NA).

index_source
(optional) Character value: if "BOA" (default), indices are computed from BOA values; if "TOA", non corrected reflectances are instead used (be careful to use this setting!).

rgb_ranges
(optional) Range of valid values to be used for RGB products. Values must be provided in the same scale used within SAFE and BOA/TOA products (0-10000, corresponding to reflectances * 10000). If can be a 2-length integer vector (min-max for all the 3 bands) or a 6-length vector or 3x2 matrix (min red, min green, min blue, max red, max green, max blue). Default is to use c(0,2500) for bands 2, 3 and 4; c(0,7500) for other bands. In case list_rgb is a vector of length > 1, rgb_ranges must be a list of the same length (otherwise, the same range vlaues will be used for all the RGB products).

mask_type
(optional) Character value which determines the categories in the Surface Classification Map to be masked (see s2_mask() for the accepted values). Default (NA) is not to mask.

max_mask
(optional) Numeric value (range 0 to 100), which represents the maximum percentage of allowed masked surface (by clouds or any other type of mask chosen with argument mask_type) for producing outputs. Images with a percentage of masked surface greater than max_mask are not processed (the list of expected output files which have not been generated is returned as an attribute, named "skipped"). Default value is 80. This parameter is different from max_cloud_safe, because:

it is computed over the selected extent;

it is computed basing on the cloud mask defined as above. Notice that the percentage is computed on non-NA values (if input images had previously been clipped and masked using a polygon, the percentage is computed on the surface included in the masking polygons).

mask_smooth
(optional) Numeric positive value: the smoothing radius (expressed in unit of measure of the output projection, typically metres) to be applied to the cloud mask by function s2_mask.

mask_buffer
(optional) Numeric value: the buffering radius (expressed in unit of measure of the output projection, typically metres) to be applied to the cloud mask by function s2_mask. Default value (0) means that no buffer is applied; a positive value causes an enlargement of the masked area; a negative value cause a reducement.

clip_on_extent
(optional) Logical: if TRUE (default), output products and indices are clipped to the selected extent (and resampled/reprojected); if FALSE, the geometry and extension of the tiles is maintained.

extent_as_mask
(optional) Logical: if TRUE, pixel values outside the extent polygon are set to NA; if FALSE (default), all the values within the bounding box are maintained.

reference_path
(optional) Path of the raster file to be used as a reference grid. If NA (default), no reference is used.

res
(optional) Numerifc vector of length 2 with the x-y resolution for output products. Default (NA) means that the resolution is keeped as native.

res_s2
(optional) Character value corresponding to the native Sentinel-2 resolution to be used. Accepted values are "10m" (default), "20m" and "60m".

unit
(optional) Character value corresponding to the unit of measure with which to interpret the resolution (for now, only "Meter" - the default value - is supported).

proj
(optional) Character string with the pro4string of the output resolution. default value (NA) means not to reproject.

resampling
(optional) Resampling method (one of the values supported by gdal_translate: "near" (default), "bilinear", "cubic", "cubicspline", "lanczos", "average" or "mode").

resampling_scl
(optional) Resampling method for categorical products (for now, only SCL): one among "near" (default) and "mode".

outformat
(optional) Format of the output file (in a format recognised by GDAL). Default is "GTiff".

rgb_outformat
(optional) Format of the output RGB products (in a format recognised by GDAL). Default is "GTiff".

index_datatype
(optional) Numeric datatype of the ouptut spectral indices (see s2_calcindices.

compression
(optional) In the case GTiff is chosen as output format, the compression indicated with this parameter is used (default is "DEFLATE").

rgb_compression
(optional) In the case GTiff is chosen as output format for RGB products, the compression indicated with this parameter is used (default is "DEFLATE"). In the cases GTiff or JPEG are chosen as output format for RGB products, this parameter can also be a 1-100 integer value, which is interpreted as the compression level for a JPEG compression.

overwrite
(optional) Logical value: should existing output files be overwritten? (default: FALSE).

path_l1c
(optional) Path of the directory in which Level-1C SAFE products are searched and/or downloaded. If not provided (default), a temporary directory is used.

path_l2a
(optional) Path of the directory in which Level-2A SAFE products are searched, downloaded and/or generated. If not provided (default), a temporary directory is used.

path_tiles
(optional) Path of the directory in which Sentinel-2 tiles (as generated by s2_translate) are searched and/or generated. If not provided (default), a temporary directory is used, and files are generated as virtual rasters; otherwise, they are generated in the format specified with outformat parameter.

path_merged
(optional) Path of the directory in which Sentinel-2 tiles merged by orbit (as generated by s2_merge) are searched and/or generated. If not provided (default), a temporary directory is used, and files are generated as virtual rasters; otherwise, they are generated in the format specified with outformat parameter.

path_out
(optional) Path of the directory in which Sentinel-2 output products are searched and/or generated. If not provided (default), a temporary directory is used.

path_rgb
(optional) Path of the directory in RGB products are searched and/or generated. If not provided (default), path_out is used.

path_indices
(optional) Path of the directory in which files of spectral indices are searched and/or generated. If not provided (default), path_out is used.

path_subdirs
(optional) Logical: if TRUE (default), a directory for each output product or spectral index is generated within path_tiles, path_merged, path_out and path_indices; if FALSE, products are put directly within them.

thumbnails
(optional) Logical: if TRUE (default), a thumbnail is added for each product created. Thumbnails are JPEG or PNG georeferenced small images (width or height of 1024 pixels) with default colour palettes (for more details, see the help window in the GUI). They are placed in a subdirectory of the products names "thumbnails". If FALSE, they are not created.

parallel
(optional) Logical or integer: if TRUE (default), some functions (sen2cor, s2_mask and s2_calcindices for now) are executed using multiple cores in order to speed up the execution. The number of cores is automatically determined; specifying it is also possible (e.g. parallel = 4). If FALSE, the processing chain is forced to run with a single core (this can be useful if multiple sen2r instances are run in parallel). This argument can be set only in commandline mode, not using the GUI.

use_python
(optional) Logical: if TRUE (default), the presence of python in the system is checked before running the function; if FALSE, this is skipped. Setting this to FALSE can bge useful on systems with problems with python, when sen2r() is intended to be used only for processing existing SAFE files (python is required in any case to download SAFE).

tmpdir
(optional) Path where intermediate files will be created. Default is a temporary directory (unless outformat = "VRT": in this case, default is a subdirectory named ".vrt" within path_out).

rmtmp
(optional) Logical: should temporary files be removed? (Default: TRUE). rmtmp is forced to FALSE if outformat = "VRT".

log
(optional) Character string with the path where the package messages will be redirected. Default (NA) is not to redirect (use standard output). A two-length character with tho paths (which can also coincide) can be used to redirect also the output: in this case, the first path is the path for messages, the second one for the output.

Value

A vector with the paths of the files which were created (excluded the temporary files); NULL otherwise. The vector includes two attributes:

cloudcovered with the list of imags not created due to the higher percentage of cloud covered pixels;

missing with the list of imags not created due to other reasons.