Summarise processing report

Internal function used to summarise results of sen2r execution at the end of a processing..

sen2r_process_report(
  s2_list_ordered,
  s2names = NULL,
  pm = NULL,
  ignorelist = list(),
  s2_list_cloudcovered = NA,
  s2_list_failed = NA,
  download_only = FALSE,
  s2_downloaded = NA,
  s2_skipped = NA,
  s2_corrected = NA
)

Arguments

s2_list_ordered
List containing the lists of ordered/notordered LTA S2 images.

s2names
Output of compute_s2_paths().

pm
List containing sen2r processing parameters.

ignorelist
Internal parameter.

s2_list_cloudcovered
Internal parameter.

s2_list_failed
Internal parameter.

download_only
Logical: if TRUE, it indicates that the processing to be summarised only involved download (pm$preprocess = FALSE).

s2_downloaded
Internal parameter.

s2_skipped
Internal parameter.

s2_corrected
Internal parameter.

Value

A data.frame summarising the report, and containing the following columns:

time: date/time of report creation;

n_req_tot_dates: number of dates to be processed based on the query;

n_ondisk_dates: number of dates for which all expected products are already on disk;

n_proc_dates: number of date for which some products were computed in the current run;

n_complete_out: number of dates for which processing is "complete", and all products were created;

n_failed_dates: number of dates for which processing is "complete", but some products were not created for unexpected reasons;

n_cloudy_dates: number of dates for which processing is "complete", but some products were not created because cloudiness in the spatial extent was above max_mask;

n_notonline_dates: number of date for which processing is "incomplete", because not all require images are online;

n_ordered_imgs: number of images correctly ordered from LTA;

n_notordered_imgs: number of images for which LTA order failed;

n_downloaded: number of images downloaded during current run;

n_skipped: number of required images not downloaded because already on disk (note: this does not include images that would be needed to process a date for which all products are already on disk);

n_corrected: number of images atmospherically corrected using sen2cor;

completed: logical, indicating if processing can be considered "complete" (it is set to TRUE in case n_notonline_dates = 0).

Note

License: GPL 3.0

References

L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020). "sen2r": An R toolbox for automatically downloading and preprocessing Sentinel-2 satellite data. Computers & Geosciences, 139, 104473. doi: 10.1016/j.cageo.2020.104473 , URL: http://sen2r.ranghetti.info/.

Author

Lorenzo Busetto, phD (2020)