Return WKT or WKT2 basing on the installed rgdal version

This is a convenience temporary function which returns the WKT representation of a CRS, using sf::st_as_text in case PROJ < 3, rgdal::CRS otherwise. This has the advantage to perform precise transformations with PROJ >=3, and to avoid conversion errors (see here). This function will be deleted whenever sf will manage WKT2.

st_as_text_2(x, pretty = FALSE)

Arguments

x
object of class sfg, sfc or crs

pretty
logical; if TRUE, print human-readable well-known-text representation of a coordinate reference system

Value

Well-known Text representation of simple feature geometry or coordinate reference system

Examples

sen2r:::st_as_text_2(sf::st_crs(32632))

#> [1] "PROJCS[\"UTM Zone 32, Northern Hemisphere\",GEOGCS[\"WGS 84\",DATUM[\"WGS_1984\",SPHEROID[\"WGS 84\",6378137,298.257223563,AUTHORITY[\"EPSG\",\"7030\"]],AUTHORITY[\"EPSG\",\"6326\"]],PRIMEM[\"Greenwich\",0,AUTHORITY[\"EPSG\",\"8901\"]],UNIT[\"degree\",0.0174532925199433,AUTHORITY[\"EPSG\",\"9122\"]],AUTHORITY[\"EPSG\",\"4326\"]],PROJECTION[\"Transverse_Mercator\"],PARAMETER[\"latitude_of_origin\",0],PARAMETER[\"central_meridian\",9],PARAMETER[\"scale_factor\",0.9996],PARAMETER[\"false_easting\",500000],PARAMETER[\"false_northing\",0],UNIT[\"Meter\",1]]"