Compute names of S2 file to be generated

Internal function (to be used within sen2r()) which computes the names of the required output image files (see details). The function was splitted from sen2r() because this code is called twice (and to shorten the main function).

compute_s2_paths(pm, s2_list_l1c, s2_list_l2a, paths, list_prods, out_ext,
  index_ext, tiles_ext, merged_ext, warped_ext, sr_masked_ext,
  force_tiles = FALSE, check_tmp = TRUE, ignorelist)

Arguments

pm
List of input parameters.

s2_list_l1c
Names and paths of input SAFE level-1C products.

s2_list_l2a
Names and paths of input SAFE level-2A products.

paths
Named vector of required paths. It must contain elements named "out", "indices", "tiles", "merged" and "warped".

list_prods
Character vector with the values of the products to be processed (accepted values: "TOA", "BOA", "SCL", "TCI").

out_ext
Extension (character) of output products.

index_ext
Extension (character) of index products.

tiles_ext
Extension (character) of tiled products.

merged_ext
Extension (character) of merged products.

warped_ext
Extension (character) of warped products.

sr_masked_ext
Extension (character) of masked products of SR products.

force_tiles
(optional) Logical: passed to s2_shortname (default: FALSE).

check_tmp
(optional) Logical: if TRUE (default), temporary files are also searched when _exi names are computed; if FALSE, only non temporary files are searched.

ignorelist
Vector of output files to be ignored.

Value

A list, containing the following vectors: tiles_names_exp, merged_names_exp, warped_names_exp, masked_names_exp, out_names_exp, indices_names_exp, indices_names_new, out_names_req, out_names_new, masked_names_new, warped_names_req, warped_names_new, merged_names_req, merged_names_new, tiles_names_req, tiles_names_new, safe_names_l1c_req, safe_names_l2a_req.

Details

This fuction is structured in the following way:

Retrieve the file names expected to be present at the end of the processing chain (suffix _exp):

tiles_names_

merged_names_

warped_names_

masked_names_

out_names_ (= warped_names_ or subset(warped_names_)+masked_names_)

indices_names_

Compute the file names expected to be created (suffixes _req and _new, see below) (this operation is done in reverse order).

Meaning of the suffixes _exp, _req and _new (here and for all the script):

_exp: full names of the files expected to be present at the end of the processing chain (already existing or not);

_req: names of the files required for the next step (e.g. tiles_names_req are required to perform s2_merge())

_new: names of the required files not existing yet (expected to be created).

With overwrite=TRUE, all these vectors are equal (e.g. merged_names_exp = merged_names_req = merged_names_new), because all is overwritten.

Note

License: GPL 3.0