Preprocess (clip, reproject, reshape, mask) SAFE products

The function is a simplified wrapper of sen2r() to preprocess existing SAFE products. No online research is performed, neither sen2cor is applied to existing L1C products. Output products are GeoTIFF files clipped to a desired

geograbber_process(path_safe, path_out, bbox = NA, list_prods = c("BOA"),
  mask_type = NA, res = NA, proj = NA, resampling = "near",
  overwrite = FALSE)

Arguments

path_safe
Path of the directory in which SAFE products are searched.

path_out
Path of the directory in which Sentinel-2 output products are generated.

bbox
(optional) Four-length numeric vector with the bounding box of the output products (xmin, ymin, xmax, ymax), in geographic coordinates.

list_prods
(optional) Character vector with the values of the products to be processed (accepted values: "TOA", "BOA", "SCL", "TCI"). Default is "BOA".

mask_type
(optional) Character value which determines the categories in the Srface Classification Map to be masked (see s2_mask() for the accepted values). Default (NA) is not to mask.

res
(optional) Numerifc vector of length 2 with the x-y resolution for output products. NA (default) means that the resolution is keeped as native.

proj
(optional) Character string with the pro4string of the output resolution. default value (NA) means not to reproject.

resampling
(optional) Resampling method (one of the values supported by gdal_translate: "near" (default), "bilinear", "cubic", "cubicspline", "lanczos", "average" or "mode").

overwrite
(optional) Logical value: should existing output files be overwritten? (default: FALSE).