List spectral indices

Return a table with attributes of the spectral indices computable with the package.

list_indices(values, pattern = "")

Arguments

values
A vector of attributes which will be returned, being one or more within the followings:

n_index: internal index identifiers;

name: index name;

longname: index description;

link: URL to the index description page;

s2_formula: expression containing the formula to compute the index;

s2_formula_mathml: MathML version of the formula;

checked: logical (TRUE for verified indices);

a, b, x: parameter values (NA for non required parameters).

pattern
A regular expression on index names.

Value

A data.frame with the required information. The table contains also the following attributes:

creation_date: timestamp of the creation date of the indices archive;

pkg_version: version of the sen2r package used to create the indices archive.

Note

License: GPL 3.0

Examples

# NOT RUN {
list_indices(c("name","longname"))
# }