List spectral indices

Return a table with attributes of the spectral indices computable with the package.

list_indices(values, pattern = "", all = FALSE)

Arguments

values
A vector of attributes which will be returned, being one or more within the followings:

n_index: internal index identifiers;

name: index name;

longname: index description;

link: URL to the index description page;

s2_formula: expression containing the formula to compute the index;

s2_formula_mathml: MathML version of the formula;

checked: logical (TRUE for verified indices);

a, b, x: parameter values (NA for non required parameters).

pattern
A regular expression on index names.

all
Logical: if TRUE, all the indices retrieved from IDB are returned; if FALSE (default), only indices checked by the authors are returned.

Value

A data.frame with the required information. The table contains also the following attributes:

creation_date: timestamp of the creation date of the indices archive;

pkg_version: version of the sen2r package used to create the indices archive.

Note

License: GPL 3.0

Examples

# Show index names
list_indices(c("name","longname"))

#>            name
#> 4           ARI
#> 6          ARVI
#> 14  Chlred-edge
#> 26          EVI
#> 34        GNDVI
#> 43  MCARI-MTVI2
#> 51        MCARI
#> 53       MCARI2
#> 58       MSAVI2
#> 61        MTVI2
#> 78         NDII
#> 88        BNDVI
#> 90         NDVI
#> 91         NDRE
#> 92          NBR
#> 96        OSAVI
#> 108         MSI
#> 133      PSSRb1
#> 181        SAVI
#> 186       SIPI1
#> 195 TCARI-OSAVI
#> 196       TCARI
#> 199       TCIdx
#> 205        NDFI
#> 206       NDFI2
#> 207        NDSI
#> 208        NBR2
#> 209       MIRBI
#> 210         CSI
#> 211       CRred
#> 215       BDred
#> 223        NDWI
#> 224       NDWI2
#>                                                                                         longname
#> 4                                                                  Anthocyanin reflectance index
#> 6                                                     Atmospherically Resistant Vegetation Index
#> 14                                                                          Chlorophyll Red-Edge
#> 26                                                                     Enhanced Vegetation Index
#> 34                                                  Green Normalized Difference Vegetation Index
#> 43                                                                                   MCARI/MTVI2
#> 51                                          Modified Chlorophyll Absorption in Reflectance Index
#> 53                                        Modified Chlorophyll Absorption in Reflectance Index 2
#> 58                                                       Modified Soil Adjusted Vegetation Index
#> 61                                                        Modified Triangular Vegetation Index 2
#> 78                                                           Normalized Difference 819/1600 NDII
#> 88                    Normalized Difference NIR/Blue Blue-normalized difference vegetation index
#> 90  Normalized Difference NIR/Red Normalized Difference Vegetation Index, Calibrated NDVI - CDVI
#> 91                              Normalized Difference NIR/Rededge Normalized Difference Red-Edge
#> 92                                          Normalized Difference NIR/SWIR Normalized Burn Ratio
#> 96                                                      Optimized Soil Adjusted Vegetation Index
#> 108                                                  Simple Ratio 1600/820 Moisture Stress Index
#> 133                                        Simple Ratio 800/650 Pigment specific simple ratio B1
#> 181                                                               Soil Adjusted Vegetation Index
#> 186                                                          Structure Intensive Pigment Index 1
#> 195                                                                                  TCARI/OSAVI
#> 196                                                     Transformed Chlorophyll Absorbtion Ratio
#> 199                                                                 Triangular chlorophyll index
#> 205                                                       Normalized Difference Flood Index B1B7
#> 206                                                       Normalized Difference Flood Index B1B6
#> 207                                                              Normalize Difference Snow Index
#> 208                                                                      Normalized Burn Ratio 2
#> 209                                                                      Mid-Infrared Burn Index
#> 210                                                                              Char Soil Index
#> 211                                                                 Continuum Removal in the red
#> 215                                                                        Band Depth in the red
#> 223                                                            Normalized Difference Water Index
#> 224                                                          Normalized Difference Water Index 2

# Return the MSAVI2 formula
list_indices("s2_formula", "^MSAVI2$")

#> [1] "(2*band_8+1-sqrt(power((2*band_8+1),2)-8*(band_8-band_4)))/(2)"
#> attr(,"pkg_version")
#> [1] ‘1.2.0’
#> attr(,"creation_date")
#> [1] "2019-10-11 17:16:25 CEST"

# Return all index names (including unchecked)
list_indices("name", all = TRUE)

#>   [1] "ATSAVI"        "AFRI1600"      "AFRI2100"      "ARI"          
#>   [5] "AVI"           "ARVI"          "ARVI2"         "BWDRVI"       
#>   [9] "BRI"           "CARI"          "Chlgreen"      "CIgreen"      
#>  [13] "CIrededge"     "Chlred-edge"   "CVI"           "CI"           
#>  [17] "Datt1"         "Datt4"         "Datt6"         "D678-500"     
#>  [21] "D800-550"      "D800-680"      "D833-658"      "GDVI"         
#>  [25] "DVIMSS"        "EVI"           "EVI2"          "EPI"          
#>  [29] "Fe2+"          "Fe3+"          "GVMI"          "GARI"         
#>  [33] "GLI"           "GNDVI"         "GOSAVI"        "GSAVI"        
#>  [37] "GBNDVI"        "GRNDVI"        "IVI"           "IPVI"         
#>  [41] "I"             "LCI"           "MCARI-MTVI2"   "MCARI-OSAVI"  
#>  [45] "MVI"           "MGVI"          "MNSI"          "MSBI"         
#>  [49] "MYVI"          "mND680"        "MCARI"         "MCARI1"       
#>  [53] "MCARI2"        "mNDVI"         "mSR"           "MSR670"       
#>  [57] "MSRNir-Red"    "MSAVI2"        "MSAVIhyper"    "MTVI1"        
#>  [61] "MTVI2"         "Norm G"        "Norm NIR"      "Norm R"       
#>  [65] "PPR"           "PVR"           "ND774-677"     "GNDVIhyper"   
#>  [69] "ND782-666"     "ND790-670"     "ND800-2170"    "PSNDc2"       
#>  [73] "PSNDc1"        "GNDVIhyper2"   "PSNDb1"        "PSNDa1"       
#>  [77] "ND800-680"     "NDII"          "NDII2"         "NDMI"         
#>  [81] "ND827-668"     "ND833-1649"    "ND833-658"     "SIWSI"        
#>  [85] "ND895-675"     "NGRDI"         "NDVI2"         "BNDVI"        
#>  [89] "MNDVI"         "NDVI"          "NDRE"          "NBR"          
#>  [93] "RI"            "NDSaI"         "NDVI690-710"   "OSAVI"        
#>  [97] "PNDVI"         "PVI"           "Rededge1"      "Rededge2"     
#> [101] "RBNDVI"        "REP"           "Rre"           "RDVI"         
#> [105] "SAVImir"       "IF"            "MSI2"          "MSI"          
#> [109] "TM5-TM7"       "SR440-740"     "BGI"           "SR520-670"    
#> [113] "SR550-670"     "DSWI-4"        "SR550-800"     "GI"           
#> [117] "SR560-658"     "SR672-550"     "SR672-708"     "SR674-553"    
#> [121] "SR675-555"     "SR675-700"     "SR675-705"     "SR700"        
#> [125] "SR700-670"     "SR710-670"     "SR735-710"     "SR774-677"    
#> [129] "SR800-2170"    "PSSRc2"        "PSSRc1"        "SR800-550"    
#> [133] "PSSRb1"        "RVI"           "PSSRa1"        "SR800-680"    
#> [137] "SR801-550"     "SR801-670"     "PBI"           "SR833-1649"   
#> [141] "SR833-658"     "Datt2"         "SR860-550"     "SR860-708"    
#> [145] "RDI"           "SRMIR-Red"     "SRNir-700-715" "GRVI"         
#> [149] "SRNIR-MIR"     "DVI"           "RRI1"          "IO"           
#> [153] "RGR"           "SRRed-NIR"     "SB1580"        "SB2100"       
#> [157] "SB2130"        "SB2180"        "SB2218"        "SB2240"       
#> [161] "SB2250"        "SB2270"        "SB2280"        "SB460"        
#> [165] "BB3"           "SR495"         "SB550"         "SB555"        
#> [169] "SB655"         "SB660"         "SB670"         "SB675"        
#> [173] "BB1"           "SB700"         "SB703"         "SB705"        
#> [177] "SB735"         "SB801"         "SB850"         "SB885"        
#> [181] "SAVI"          "SBL"           "SAVI2"         "SLAVI"        
#> [185] "SQRT(IR-R)"    "SIPI1"         "SIPI3"         "SBI"          
#> [189] "GVIMSS"        "NSIMSS"        "SBIMSS"        "GVI"          
#> [193] "WET"           "YVIMSS"        "TCARI-OSAVI"   "TCARI"        
#> [197] "TSAVI"         "TSAVI2"        "TCIdx"         "VI700"        
#> [201] "VARIgreen"     "VARIrededge"   "WDVI"          "WDRVI"        
#> [205] "NDFI"          "NDFI2"         "NDSI"          "NBR2"         
#> [209] "MIRBI"         "CSI"           "CRred"         "CRred-2A"     
#> [213] "CRred-2B"      "CRred-0"       "BDred"         "BDred-2A"     
#> [217] "BDred-2B"      "BDred-0"       "CRred2"        "CRred2-2A"    
#> [221] "CRred2-2B"     "CRred2-0"      "NDWI"          "NDWI2"        
#> attr(,"pkg_version")
#> [1] ‘1.2.0’
#> attr(,"creation_date")
#> [1] "2019-10-11 17:16:25 CEST"