Get metadata from raster paths

This accessory function extract some useful metadata from a vector of raster paths.

raster_metadata(raster_paths, meta = "all", format = "data.table")

Arguments

raster_paths
A vector of raster paths.

meta
Vector with the desired metadata: one or more values among 'res', 'size', 'bbox', 'proj', 'unit', 'outformat', 'type'. Alternatively meta = 'all' (default) allows to return all metadata.

format
One between data.table (default), data.frame and list.

Value

A data.table, data.frame or list of the output metadata.

Note

License: GPL 3.0

References

L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020). "sen2r": An R toolbox for automatically downloading and preprocessing Sentinel-2 satellite data. Computers & Geosciences, 139, 104473. DOI: 10.1016/j.cageo.2020.104473, URL: http://sen2r.ranghetti.info/.

Examples

# Define product names
examplenames <- c(
  system.file("tif/L7_ETMs.tif", package="stars"),
  system.file("extdata/out/S2A2A_20190723_022_Barbellino_BOA_10.tif",
    package = "sen2r")
)

# \donttest{
# Return metadata as data.table
raster_metadata(examplenames)

#>                                                                                                           path
#> 1:                                    /home/lranghetti/R/x86_64-pc-linux-gnu-library/3.6/stars/tif/L7_ETMs.tif
#> 2: /home/lranghetti/share/git/github/ranghetti/sen2r/inst/extdata/out/S2A2A_20190723_022_Barbellino_BOA_10.tif
#>    valid res.x res.y size.x size.y nbands     xmin    ymin     xmax    ymax
#> 1:  TRUE  28.5  28.5    349    352      6 288776.3 9110729 298722.8 9120761
#> 2:  TRUE  10.0  10.0     24     42     11 580560.0 5101700 580800.0 5102120
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                        proj
#> 1: PROJCS["UTM Zone 25, Southern Hemisphere",GEOGCS["GRS 1980(IUGG, 1980)",DATUM["unknown",SPHEROID["GRS80",6378137,298.257222101],TOWGS84[0,0,0,0,0,0,0]],PRIMEM["Greenwich",0],UNIT["degree",0.0174532925199433]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",-33],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",10000000],UNIT["metre",1,AUTHORITY["EPSG","9001"]]]
#> 2:                                                                                                                                                                                                                                                                                                                                                                                                                                                               EPSG:32632
#>     unit outformat   type
#> 1: metre     GTiff   Byte
#> 2: metre     GTiff UInt16

# }

# Return some metadata as data.table
raster_metadata(examplenames, c("res", "size", "bbox", "outformat"))

#>                                                                                                           path
#> 1:                                    /home/lranghetti/R/x86_64-pc-linux-gnu-library/3.6/stars/tif/L7_ETMs.tif
#> 2: /home/lranghetti/share/git/github/ranghetti/sen2r/inst/extdata/out/S2A2A_20190723_022_Barbellino_BOA_10.tif
#>    valid res.x res.y size.x size.y     xmin    ymin     xmax    ymax outformat
#> 1:  TRUE  28.5  28.5    349    352 288776.3 9110729 298722.8 9120761     GTiff
#> 2:  TRUE  10.0  10.0     24     42 580560.0 5101700 580800.0 5102120     GTiff

# Return some metadata as list
raster_metadata(examplenames, c("res", "size", "bbox", "proj"), format = "list")

#> [[1]]
#> [[1]]$path
#> [1] "/home/lranghetti/R/x86_64-pc-linux-gnu-library/3.6/stars/tif/L7_ETMs.tif"
#> 
#> [[1]]$valid
#> [1] TRUE
#> 
#> [[1]]$res
#>    x    y 
#> 28.5 28.5 
#> 
#> [[1]]$size
#>   x   y 
#> 349 352 
#> 
#> [[1]]$bbox
#>      xmin      ymin      xmax      ymax 
#>  288776.3 9110728.8  298722.8 9120760.8 
#> 
#> [[1]]$proj
#> Coordinate Reference System:
#>   User input: PROJCS["UTM Zone 25, Southern Hemisphere",GEOGCS["GRS 1980(IUGG, 1980)",DATUM["unknown",SPHEROID["GRS80",6378137,298.257222101],TOWGS84[0,0,0,0,0,0,0]],PRIMEM["Greenwich",0],UNIT["degree",0.0174532925199433]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",-33],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",10000000],UNIT["metre",1,AUTHORITY["EPSG","9001"]]] 
#>   wkt:
#> PROJCS["UTM Zone 25, Southern Hemisphere",
#>     GEOGCS["GRS 1980(IUGG, 1980)",
#>         DATUM["unknown",
#>             SPHEROID["GRS80",6378137,298.257222101],
#>             TOWGS84[0,0,0,0,0,0,0]],
#>         PRIMEM["Greenwich",0],
#>         UNIT["degree",0.0174532925199433]],
#>     PROJECTION["Transverse_Mercator"],
#>     PARAMETER["latitude_of_origin",0],
#>     PARAMETER["central_meridian",-33],
#>     PARAMETER["scale_factor",0.9996],
#>     PARAMETER["false_easting",500000],
#>     PARAMETER["false_northing",10000000],
#>     UNIT["metre",1,
#>         AUTHORITY["EPSG","9001"]]]
#> 
#> 
#> [[2]]
#> [[2]]$path
#> [1] "/home/lranghetti/share/git/github/ranghetti/sen2r/inst/extdata/out/S2A2A_20190723_022_Barbellino_BOA_10.tif"
#> 
#> [[2]]$valid
#> [1] TRUE
#> 
#> [[2]]$res
#>  x  y 
#> 10 10 
#> 
#> [[2]]$size
#>  x  y 
#> 24 42 
#> 
#> [[2]]$bbox
#>    xmin    ymin    xmax    ymax 
#>  580560 5101700  580800 5102120 
#> 
#> [[2]]$proj
#> Coordinate Reference System:
#>   User input: PROJCS["WGS 84 / UTM zone 32N",GEOGCS["WGS 84",DATUM["WGS_1984",SPHEROID["WGS 84",6378137,298.257223563,AUTHORITY["EPSG","7030"]],AUTHORITY["EPSG","6326"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4326"]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",9],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",0],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","32632"]] 
#>   wkt:
#> PROJCS["WGS 84 / UTM zone 32N",
#>     GEOGCS["WGS 84",
#>         DATUM["WGS_1984",
#>             SPHEROID["WGS 84",6378137,298.257223563,
#>                 AUTHORITY["EPSG","7030"]],
#>             AUTHORITY["EPSG","6326"]],
#>         PRIMEM["Greenwich",0,
#>             AUTHORITY["EPSG","8901"]],
#>         UNIT["degree",0.0174532925199433,
#>             AUTHORITY["EPSG","9122"]],
#>         AUTHORITY["EPSG","4326"]],
#>     PROJECTION["Transverse_Mercator"],
#>     PARAMETER["latitude_of_origin",0],
#>     PARAMETER["central_meridian",9],
#>     PARAMETER["scale_factor",0.9996],
#>     PARAMETER["false_easting",500000],
#>     PARAMETER["false_northing",0],
#>     UNIT["metre",1,
#>         AUTHORITY["EPSG","9001"]],
#>     AXIS["Easting",EAST],
#>     AXIS["Northing",NORTH],
#>     AUTHORITY["EPSG","32632"]]
#> 
#> 

# Output with an invalid raster
examplenames <- c(
  examplenames,
  system.file("extdata/settings/gdal_formats.json", package="sen2r")
)
raster_metadata(examplenames, c("bbox", "proj"))

#> trying to read file: /home/lranghetti/share/git/github/ranghetti/sen2r/inst/extdata/settings/gdal_formats.json

#>                                                                                                           path
#> 1:                                    /home/lranghetti/R/x86_64-pc-linux-gnu-library/3.6/stars/tif/L7_ETMs.tif
#> 2: /home/lranghetti/share/git/github/ranghetti/sen2r/inst/extdata/out/S2A2A_20190723_022_Barbellino_BOA_10.tif
#> 3:                   /home/lranghetti/share/git/github/ranghetti/sen2r/inst/extdata/settings/gdal_formats.json
#>    valid     xmin    ymin     xmax    ymax
#> 1:  TRUE 288776.3 9110729 298722.8 9120761
#> 2:  TRUE 580560.0 5101700 580800.0 5102120
#> 3: FALSE       NA      NA       NA      NA
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                        proj
#> 1: PROJCS["UTM Zone 25, Southern Hemisphere",GEOGCS["GRS 1980(IUGG, 1980)",DATUM["unknown",SPHEROID["GRS80",6378137,298.257222101],TOWGS84[0,0,0,0,0,0,0]],PRIMEM["Greenwich",0],UNIT["degree",0.0174532925199433]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",-33],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",10000000],UNIT["metre",1,AUTHORITY["EPSG","9001"]]]
#> 2:                                                                                                                                                                                                                                                                                                                                                                                                                                                               EPSG:32632
#> 3:                                                                                                                                                                                                                                                                                                                                                                                                                                                                     <NA>