Retrieve list of available S2 products.

The function retrieves the list of available Sentinel-2 products satisfying given search criteria.

s2_list(spatial_extent = NULL, tile = NULL, orbit = NULL,
  time_interval = c(Sys.Date() - 10, Sys.Date()), time_period = "full",
  level = "auto", apihub = NA, max_cloud = 100,
  output_type = "vector")

Arguments

spatial_extent
A valid spatial object object of class sf, sfc or sfg

tile
string array Sentinel-2 Tiles to be considered string (5-length character)

orbit
string array Sentinel-2 orbit numbers to be considered

time_interval
Dates to be considered, as a temporal vector (class POSIXct or Date, or string in YYYY-mm-dd format) of length 1 (specific day) or 2 (time interval).

time_period
(optional) Character:

"full" (default) means that all the images included in the time window are considered;

"seasonal" means that only the single seasonal periods in the window are used (i.e., with a time window from 2015-06-01 to 2017-08-31, the periods 2015-06-01 to 2015-08-31, 2016-06-01 to 2016-08-31 and 2017-06-01 to 2017-08-31 are considered).

level
Character vector with one of the following: - "auto" (default): check if level-2A is available on SciHub: if so, list it; if not, list the corresponding level-1C product - "L1C": list available level-1C products - "L2A": list available level-2A products

apihub
Path of the "apihub.txt" file containing credentials of SciHub account. If NA (default), the default location inside the package will be used.

max_cloud
Integer number (0-100) containing the maximum cloud level of the tiles to be listed (default: no filter).

output_type
Character: if 'vector' (default), the function returns a vector or URLs, whose names are the SAFE names; if 'data.table', the output is a data.table with metadata.

Value

Depending on the value of argument `output_type``, a vector of available products (being each element an URL and its name the product name), or a data.table with product metadata.

Note

License: GPL 3.0

Examples

# \donttest{
pos <- sf::st_sfc(sf::st_point(c(9.85,45.81)), crs = 4326)
time_window <- as.Date(c("2016-05-01", "2017-07-30"))

# Full-period list
example_s2_list <- s2_list(
  spatial_extent = pos,
  tile = "32TNR",
  time_interval = time_window,
  orbit = "065"
)
print(example_s2_list)

#>                                      S2A_MSIL1C_20160502T102032_N0201_R065_T32TNR_20160502T102027.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('fde01d3f-8a0d-46cb-b65d-3e32eff93f1d')/$value" 
#>                                      S2A_MSIL1C_20160512T102032_N0202_R065_T32TNR_20160512T102029.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('4b785773-ba94-49b2-ae81-9b4b08749f70')/$value" 
#>                                      S2A_MSIL1C_20160522T102032_N0202_R065_T32TNR_20160522T102029.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('4abf7965-0b26-431b-8708-74de427c0112')/$value" 
#>                                      S2A_MSIL1C_20160611T102032_N0202_R065_T32TNR_20160611T102026.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('68afcf6b-4ec6-48da-9b65-6cd41ff47a49')/$value" 
#>                                      S2A_MSIL1C_20160621T102022_N0204_R065_T32TNR_20160621T102024.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('9bb6d780-c226-4f59-bad0-1daedfdce9d5')/$value" 
#>                                      S2A_MSIL1C_20160701T102022_N0204_R065_T32TNR_20160701T102057.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('c6577b64-789e-4f7b-ba9e-a58dc4248462')/$value" 
#>                                      S2A_MSIL1C_20160711T102032_N0204_R065_T32TNR_20160711T102030.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('b61a343c-7f2d-4852-9455-09ee9f31ac35')/$value" 
#>                                      S2A_MSIL1C_20160721T102032_N0204_R065_T32TNR_20160721T102059.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('694e76ee-308e-4497-8016-87c23974247b')/$value" 
#>                                      S2A_MSIL1C_20160731T102032_N0204_R065_T32TNR_20160731T102107.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('b7a5f55b-a17e-4fc0-a4a8-9f17f1224df5')/$value" 
#>                                      S2A_MSIL1C_20160810T102032_N0204_R065_T32TNR_20160810T102044.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('a0508149-b79b-4792-8ead-00784847120b')/$value" 
#>                                      S2A_MSIL1C_20160820T102022_N0204_R065_T32TNR_20160820T102112.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('1319a17f-8611-4d3d-b4fd-1d38180625b0')/$value" 
#>                                      S2A_MSIL1C_20160830T102022_N0204_R065_T32TNR_20160830T102052.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('848b6310-f11d-407c-acff-9f48bd76d73e')/$value" 
#>                                      S2A_MSIL1C_20160909T102022_N0204_R065_T32TNR_20160909T102108.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('846a0222-321c-4caf-8606-40f4719b5356')/$value" 
#>                                      S2A_MSIL1C_20160919T102022_N0204_R065_T32TNR_20160919T102112.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('e7f860fa-8534-4bb1-aaf5-85271669bf8d')/$value" 
#>                                      S2A_MSIL1C_20160929T102022_N0204_R065_T32TNR_20160929T102344.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('38fd8601-6b61-4079-9007-21aa5d0659a4')/$value" 
#>                                      S2A_MSIL1C_20161009T102022_N0204_R065_T32TNR_20161009T102522.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('326c2d14-8da3-47a0-86e5-abbc993c22c4')/$value" 
#>                                      S2A_MSIL1C_20161019T102032_N0204_R065_T32TNR_20161019T102353.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('02aea734-113c-47e9-9edc-7bb8002b17a1')/$value" 
#>                                      S2A_MSIL1C_20161029T102132_N0204_R065_T32TNR_20161029T102445.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('6f6edccf-457e-489d-8b5b-8c7e50238905')/$value" 
#>                                      S2A_MSIL1C_20161108T102232_N0204_R065_T32TNR_20161108T102425.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('dc94e0b3-942c-4448-b6cc-1284f8c62705')/$value" 
#>                                      S2A_MSIL1C_20161118T102322_N0204_R065_T32TNR_20161118T102318.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('f42fd006-841d-442c-8f02-d02bcc9b294f')/$value" 
#>                                      S2A_MSIL1C_20161128T102352_N0204_R065_T32TNR_20161128T102354.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('e0610a2c-f257-42de-8914-40a0636c18bd')/$value" 
#>                                      S2A_MSIL1C_20161208T102422_N0204_R065_T32TNR_20161208T102418.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('64e4337a-c64e-4c23-a653-4cde2d6249a1')/$value" 
#>                                      S2A_MSIL1C_20161218T102432_N0204_R065_T32TNR_20161218T102606.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('69ad2f6a-a792-460b-af8b-6c0c9f574ea7')/$value" 
#>                                      S2A_MSIL1C_20161228T102432_N0204_R065_T32TNR_20161228T102428.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('35dc5ae4-ff15-4411-9e0a-7e34c428cd99')/$value" 
#>                                      S2A_MSIL1C_20170107T102401_N0204_R065_T32TNR_20170107T102658.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('77fde160-69ff-455c-a36a-3748fc057647')/$value" 
#>                                      S2A_MSIL1C_20170127T102301_N0204_R065_T32TNR_20170127T102258.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('b4b682db-daf0-40d9-9ea7-4f25423fbc42')/$value" 
#>                                      S2A_MSIL1C_20170206T102211_N0204_R065_T32TNR_20170206T102733.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('ff5d2847-2f53-4875-8d77-d321d6202842')/$value" 
#>                                      S2A_MSIL1C_20170216T102101_N0204_R065_T32TNR_20170216T102204.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('03d18e2e-0e93-4d7f-a798-9c726d65b5b6')/$value" 
#>                                      S2A_MSIL1C_20170226T102021_N0204_R065_T32TNR_20170226T102458.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('cb1c6d1c-a6ef-44f6-9ca6-5c5d7d468676')/$value" 
#>                                      S2A_MSIL1C_20170308T102021_N0204_R065_T32TNR_20170308T102514.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('cda64177-76ca-47b2-a00c-d1606cdc204e')/$value" 
#>                                      S2A_MSIL1C_20170318T102011_N0204_R065_T32TNR_20170318T102014.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('d9a705d6-1fd7-408b-ae12-984f5261372f')/$value" 
#>                                      S2A_MSIL1C_20170328T102021_N0204_R065_T32TNR_20170328T102018.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('38b0ffe1-7970-4dc1-8fa7-4264b7c5f46b')/$value" 
#>                                      S2A_MSIL1C_20170407T102021_N0204_R065_T32TNR_20170407T102343.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('a8b2d2e7-ff89-47cb-bd65-7d906de54353')/$value" 
#>                                      S2A_MSIL1C_20170417T102021_N0204_R065_T32TNR_20170417T102334.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('f43021b5-cdae-4f51-beaa-fa458b6e3692')/$value" 
#>                                      S2A_MSIL2A_20170427T102031_N0205_R065_T32TNR_20170427T102025.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('6ad22bb6-42d3-4fcf-915e-4f4a6df1cc62')/$value" 
#>                                      S2A_MSIL2A_20170507T102031_N0205_R065_T32TNR_20170507T102319.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('a4a026c0-db7b-4ba8-9b09-53027ab0d7ab')/$value" 
#>                                      S2A_MSIL2A_20170517T102031_N0205_R065_T32TNR_20170517T102352.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('6b7bb7bb-7598-45ac-9f2c-a3bd9cdc092f')/$value" 
#>                                      S2A_MSIL2A_20170527T102031_N0205_R065_T32TNR_20170527T102301.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('868b5b3c-92ac-412a-8a04-167008a1f08f')/$value" 
#>                                      S2A_MSIL2A_20170606T102031_N0205_R065_T32TNR_20170606T102456.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('dba86700-3d0a-4f40-97d2-c5d10b702f90')/$value" 
#>                                      S2A_MSIL2A_20170616T102021_N0205_R065_T32TNR_20170616T102331.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('4bcd07d6-935c-4ff0-875b-b881255d791b')/$value" 
#>                                      S2A_MSIL2A_20170626T102021_N0205_R065_T32TNR_20170626T102321.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('812bc59b-a6a8-41a7-b9be-b6424c04c7e7')/$value" 
#>                                      S2B_MSIL1C_20170701T102029_N0205_R065_T32TNR_20170701T102338.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('eec60424-00d5-43b9-8a7e-1c14d7de05b0')/$value" 
#>                                      S2A_MSIL2A_20170706T102021_N0205_R065_T32TNR_20170706T102301.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('1b893590-7550-4b98-bead-355ad059da37')/$value" 
#>                                      S2B_MSIL1C_20170711T102029_N0205_R065_T32TNR_20170711T102309.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('14b4a092-2979-4722-b037-c385339b8453')/$value" 
#>                                      S2A_MSIL1C_20170716T102021_N0205_R065_T32TNR_20170716T102324.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('4958be4f-d88c-45c5-adab-afc7c44a1374')/$value" 
#>                                      S2B_MSIL1C_20170721T102029_N0205_R065_T32TNR_20170721T102331.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('0684761e-b7d2-4512-a2a5-ead864c57806')/$value" 
#>                                      S2A_MSIL2A_20170726T102021_N0205_R065_T32TNR_20170726T102259.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('10ee1711-c04f-442a-a878-e9ac275ab011')/$value" 

# Print the dates of the retrieved products
as.vector(sort(sapply(names(example_s2_list), function(x) {
  strftime(safe_getMetadata(x,"nameinfo")$sensing_datetime)
})))

#>  [1] "2016-05-02 12:20:32" "2016-05-12 12:20:32" "2016-05-22 12:20:32"
#>  [4] "2016-06-11 12:20:32" "2016-06-21 12:20:22" "2016-07-01 12:20:22"
#>  [7] "2016-07-11 12:20:32" "2016-07-21 12:20:32" "2016-07-31 12:20:32"
#> [10] "2016-08-10 12:20:32" "2016-08-20 12:20:22" "2016-08-30 12:20:22"
#> [13] "2016-09-09 12:20:22" "2016-09-19 12:20:22" "2016-09-29 12:20:22"
#> [16] "2016-10-09 12:20:22" "2016-10-19 12:20:32" "2016-10-29 12:21:32"
#> [19] "2016-11-08 11:22:32" "2016-11-18 11:23:22" "2016-11-28 11:23:52"
#> [22] "2016-12-08 11:24:22" "2016-12-18 11:24:32" "2016-12-28 11:24:32"
#> [25] "2017-01-07 11:24:01" "2017-01-27 11:23:01" "2017-02-06 11:22:11"
#> [28] "2017-02-16 11:21:01" "2017-02-26 11:20:21" "2017-03-08 11:20:21"
#> [31] "2017-03-18 11:20:11" "2017-03-28 12:20:21" "2017-04-07 12:20:21"
#> [34] "2017-04-17 12:20:21" "2017-04-27 12:20:31" "2017-05-07 12:20:31"
#> [37] "2017-05-17 12:20:31" "2017-05-27 12:20:31" "2017-06-06 12:20:31"
#> [40] "2017-06-16 12:20:21" "2017-06-26 12:20:21" "2017-07-01 12:20:29"
#> [43] "2017-07-06 12:20:21" "2017-07-11 12:20:29" "2017-07-16 12:20:21"
#> [46] "2017-07-21 12:20:29" "2017-07-26 12:20:21"

# Seasonal-period list
example_s2_list <- s2_list(
  spatial_extent = pos,
  tile = "32TNR",
  time_interval = time_window,
  time_period = "seasonal"
)
print(example_s2_list)

#>                                      S2A_MSIL1C_20160502T102032_N0201_R065_T32TNR_20160502T102027.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('fde01d3f-8a0d-46cb-b65d-3e32eff93f1d')/$value" 
#>                                      S2A_MSIL1C_20160509T101032_N0202_R022_T32TNR_20160509T101645.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('abf24c65-2ff6-4c9a-9309-425425475d36')/$value" 
#>                                      S2A_MSIL1C_20160512T102032_N0202_R065_T32TNR_20160512T102029.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('4b785773-ba94-49b2-ae81-9b4b08749f70')/$value" 
#>                                      S2A_MSIL1C_20160519T101032_N0202_R022_T32TNR_20160519T101312.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('e1f000b8-e9c3-421b-88c2-08a9046d1a82')/$value" 
#>                                      S2A_MSIL1C_20160522T102032_N0202_R065_T32TNR_20160522T102029.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('4abf7965-0b26-431b-8708-74de427c0112')/$value" 
#>                                      S2A_MSIL1C_20160529T101032_N0202_R022_T32TNR_20160529T101030.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('6b88233c-0445-4769-9af8-d7fd9c4a2567')/$value" 
#>                                      S2A_MSIL1C_20160608T101032_N0202_R022_T32TNR_20160608T101220.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('789b8971-45d1-4917-a588-fec1696a8c46')/$value" 
#>                                      S2A_MSIL1C_20160611T102032_N0202_R065_T32TNR_20160611T102026.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('68afcf6b-4ec6-48da-9b65-6cd41ff47a49')/$value" 
#>                                      S2A_MSIL1C_20160618T101032_N0204_R022_T32TNR_20160618T101515.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('e8b6a8c1-04e9-4c58-bd1e-ff93318a852a')/$value" 
#>                                      S2A_MSIL1C_20160621T102022_N0204_R065_T32TNR_20160621T102024.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('9bb6d780-c226-4f59-bad0-1daedfdce9d5')/$value" 
#>                                      S2A_MSIL1C_20160628T101032_N0204_R022_T32TNR_20160628T101026.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('f1a23872-a307-4b7f-abfb-e9785250b790')/$value" 
#>                                      S2A_MSIL1C_20160701T102022_N0204_R065_T32TNR_20160701T102057.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('c6577b64-789e-4f7b-ba9e-a58dc4248462')/$value" 
#>                                      S2A_MSIL1C_20160708T101032_N0204_R022_T32TNR_20160708T101602.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('12f38f8c-37c0-4ed3-8412-61e32e6d38ca')/$value" 
#>                                      S2A_MSIL1C_20160711T102032_N0204_R065_T32TNR_20160711T102030.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('b61a343c-7f2d-4852-9455-09ee9f31ac35')/$value" 
#>                                      S2A_MSIL1C_20160718T101032_N0204_R022_T32TNR_20160718T101028.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('fd770e3f-cb11-4a1a-90e7-077b319a0b9a')/$value" 
#>                                      S2A_MSIL1C_20160721T102032_N0204_R065_T32TNR_20160721T102059.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('694e76ee-308e-4497-8016-87c23974247b')/$value" 
#>                                      S2A_MSIL2A_20170507T102031_N0205_R065_T32TNR_20170507T102319.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('a4a026c0-db7b-4ba8-9b09-53027ab0d7ab')/$value" 
#>                                      S2A_MSIL2A_20170517T102031_N0205_R065_T32TNR_20170517T102352.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('6b7bb7bb-7598-45ac-9f2c-a3bd9cdc092f')/$value" 
#>                                      S2A_MSIL2A_20170527T102031_N0205_R065_T32TNR_20170527T102301.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('868b5b3c-92ac-412a-8a04-167008a1f08f')/$value" 
#>                                      S2A_MSIL2A_20170606T102031_N0205_R065_T32TNR_20170606T102456.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('dba86700-3d0a-4f40-97d2-c5d10b702f90')/$value" 
#>                                      S2A_MSIL2A_20170616T102021_N0205_R065_T32TNR_20170616T102331.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('4bcd07d6-935c-4ff0-875b-b881255d791b')/$value" 
#>                                      S2A_MSIL2A_20170626T102021_N0205_R065_T32TNR_20170626T102321.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('812bc59b-a6a8-41a7-b9be-b6424c04c7e7')/$value" 
#>                                      S2B_MSIL1C_20170701T102029_N0205_R065_T32TNR_20170701T102338.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('eec60424-00d5-43b9-8a7e-1c14d7de05b0')/$value" 
#>                                      S2A_MSIL2A_20170706T102021_N0205_R065_T32TNR_20170706T102301.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('1b893590-7550-4b98-bead-355ad059da37')/$value" 
#>                                      S2B_MSIL1C_20170711T102029_N0205_R065_T32TNR_20170711T102309.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('14b4a092-2979-4722-b037-c385339b8453')/$value" 
#>                                      S2A_MSIL1C_20170716T102021_N0205_R065_T32TNR_20170716T102324.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('4958be4f-d88c-45c5-adab-afc7c44a1374')/$value" 
#>                                      S2B_MSIL1C_20170721T102029_N0205_R065_T32TNR_20170721T102331.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('0684761e-b7d2-4512-a2a5-ead864c57806')/$value" 
#>                                      S2A_MSIL2A_20170726T102021_N0205_R065_T32TNR_20170726T102259.SAFE 
#> "https://scihub.copernicus.eu/apihub/odata/v1/Products('10ee1711-c04f-442a-a878-e9ac275ab011')/$value" 

# Print the dates of the retrieved products
as.vector(sort(sapply(names(example_s2_list), function(x) {
  strftime(safe_getMetadata(x,"nameinfo")$sensing_datetime)
})))

#>  [1] "2016-05-02 12:20:32" "2016-05-09 12:10:32" "2016-05-12 12:20:32"
#>  [4] "2016-05-19 12:10:32" "2016-05-22 12:20:32" "2016-05-29 12:10:32"
#>  [7] "2016-06-08 12:10:32" "2016-06-11 12:20:32" "2016-06-18 12:10:32"
#> [10] "2016-06-21 12:20:22" "2016-06-28 12:10:32" "2016-07-01 12:20:22"
#> [13] "2016-07-08 12:10:32" "2016-07-11 12:20:32" "2016-07-18 12:10:32"
#> [16] "2016-07-21 12:20:32" "2017-05-07 12:20:31" "2017-05-17 12:20:31"
#> [19] "2017-05-27 12:20:31" "2017-06-06 12:20:31" "2017-06-16 12:20:21"
#> [22] "2017-06-26 12:20:21" "2017-07-01 12:20:29" "2017-07-06 12:20:21"
#> [25] "2017-07-11 12:20:29" "2017-07-16 12:20:21" "2017-07-21 12:20:29"
#> [28] "2017-07-26 12:20:21"

# }