Merge S2 tiles with the same date and orbit

The function merge the input Sentinel-2 files with the same date, orbit number, product type and file format. Outputs are a set of products in the same format of corresponding input files.

s2_merge(infiles, outdir = ".", subdirs = NA, tmpdir = NA,
  rmtmp = TRUE, format = NA, compress = "DEFLATE",
  vrt_rel_paths = NA, out_crs = "", parallel = FALSE,
  overwrite = FALSE, .log_message = NA, .log_output = NA)

Arguments

infiles
A vector of input filenames. Input files are paths of products already converted from SAFE format to a format managed by GDAL (use s2_translate to do it); their names must be in the sen2r naming convention (safe_shortname).

outdir
(optional) Full name of the output directory where the files should be created (default: current directory). outdir can bot be an existing or non-existing directory (in the second case, its parent directory must exists). If it is a relative path, it is expanded from the common parent directory of infiles.

subdirs
(optional) Logical: if TRUE, differet output products are placed in separated outfile subdirectories; if FALSE, they are placed in outfile directory; if NA (default), subdirectories are created only if infiles relate to more than a single product.

tmpdir
(optional) Path where intermediate files (VRT) will be created. Default is a temporary directory.

rmtmp
(optional) Logical: should temporary files be removed? (Default: TRUE). This parameter takes effect only if the output files are not VRT (in this case temporary files cannot be deleted, because rasters of source bands are included within them).

format
(optional) Format of the output file (in a format recognised by GDAL). Default is to maintain each input format.

compress
(optional) In the case a GTiff format is present, the compression indicated with this parameter is used.

vrt_rel_paths
(optional) Logical: if TRUE (default on Linux), the paths present in the VRT output file are relative to the VRT position; if FALSE (default on Windows), they are absolute. This takes effect only with format = "VRT".

out_crs
(optional) proj4string (character) of the output CRS (default: the CRS of the first input file). The tiles with CRS different from out_crs will be reprojected (and a warning returned).

parallel
(optional) Logical: if TRUE, the function is run using parallel processing, to speed-up the computation for large rasters. The number of cores is automatically determined; specifying it is also possible (e.g. parallel = 4). If FALSE (default), single core processing is used.

overwrite
Logical value: should existing output files be overwritten? (default: FALSE)

.log_message
(optional) Internal parameter (it is used when the function is called by sen2r()).

.log_output
(optional) Internal parameter (it is used when the function is called by sen2r()).

Value

A vector with the names of the merged products (just created or already existing).

Note

License: GPL 3.0