Correct L1C products using sen2cor

The function uses sen2cor to manually correct L1C products. Standalone version of sen2cor 2.4.0 is used.

sen2cor(l1c_prodlist = NULL, l1c_dir = NULL, outdir = NULL,
  proc_dir = NA, tiles = NULL, parallel = FALSE, overwrite = FALSE)

Arguments

l1c_prodlist
List of L1C product names to be corrected. They can be both product names with full/relative path or only names of SAFE products (in this case, also l1c_dir argument must be provided). SAFE products must be unzipped. Note that, at this stage, all products must be in the same directory (this will be fixed).

l1c_dir
Full or relative path of input L1C products. If NULL (default), l1c_prodlist must already be a vector of full paths.

outdir
Directory where output L2A products will be placed. If NULL (default), each product is left in the parent directory of l1c_prodlist.

proc_dir
(optional) Directory where processing is applied. If NA (default), processing is done in l1c_dir and output L2A product is then moved to outdir, unless l1c_dir is a subdirectory of a SAMBA mountpoint under Linux: in this case, L1C input products are copied in a temporary directory, processing is done there and then L2A is moved to outdir. This is required under Linux systems when l1c_dir is a subdirectory of a unit mounted with SAMBA, otherwise sen2cor would produce empty L2A products.

tiles
Vector of Sentinel-2 Tile strings (5-length character) to be processed (default: process all the tiles found in the input L1C products).

parallel
(optional) Logical: if TRUE, sen2cor instances are launched in parallel using multiple cores; if FALSE (default), they are launched in series on a single core. The number of cores is automatically determined; specifying it is also possible (e.g. parallel = 4).

overwrite
Logical value: should existing output L2A products be overwritten? (default: FALSE)

Value

Vector character with the list ot the output products (being corrected or already existing)

Note

License: GPL 3.0

Examples

# NOT RUN {
pos <- sp::SpatialPoints(data.frame("x"=12.0,"y"=44.8), proj4string=sp::CRS("+init=epsg:4326"))
time_window <- as.Date(c("2017-05-01","2017-07-30"))
example_s2_list <- s2_list(spatial_extent=pos, tile="32TQQ", time_interval=time_window)
s2_download(example_s2_list, outdir=tempdir())
sen2cor(names(example_s2_list)[1], l1c_dir=tempdir(), outdir=tempdir())
# }